Enhanced caching

+++ konduit-serving-pipeline/src/main/java/ai/konduit/serving/pipeline/api/protocol/URIResolver.java
@@ -44,6 +44,36 @@ public class URIResolver {
         return false;
     }
 
+    public static void removeOutdatedCacheEntries(File metaFile) throws IOException {
+        String lifeTimeProp = System.getProperty("konduit.serving.cache.lifetime");
+        if (StringUtils.isEmpty(lifeTimeProp))
+            return;
+        final int daysTimeout = Integer.parseInt(lifeTimeProp);
+        if (daysTimeout <= 0)
+            return;
+        File tempFile = new File(cacheDirectory, "metafile.temp");
+
+        try (BufferedReader in = new BufferedReader(new FileReader(metaFile));
+            PrintWriter writer = new PrintWriter(new OutputStreamWriter(new FileOutputStream(tempFile)))) {
+            String line = StringUtils.EMPTY;
+            while((line =in.readLine())!=null) {
+                Iterable<CSVRecord> records = CSVFormat.DEFAULT.parse(new StringReader(line));
+                for (CSVRecord record : records) {
+                    long accessedTimestamp = Long.parseLong(record.get(3));
+                    long elapsedMillis = System.currentTimeMillis() - accessedTimestamp;
+                    long elapsedDays = elapsedMillis / (1000 * 60 * 60 * 24);
+                    if (elapsedDays <= daysTimeout) {
+                        writer.write(line);
+                    } else {
+                        log.info("Removing outdated cached file " + record.get(0));
+                        new File(record.get(0)).delete();
+                    }
+                }
+            }
+            FileUtils.moveFile(tempFile, metaFile);
+        }
+    }
+
     private static File cacheDirectory;
     private static File metaFile;
     static {
@@ -55,20 +85,22 @@ public class URIResolver {
         cacheDirectory = f;
         metaFile = new File(cacheDirectory, ".metadata");
         try {
-            metaFile.createNewFile();
+           if (!metaFile.exists()) {
+               metaFile.createNewFile();
+           }
+           removeOutdatedCacheEntries(metaFile);
         } catch (IOException e) {
-            log.error("Cache initialization failed", e);
+                log.error("Cache initialization failed", e);
         }
     }
-
 
     public static File getCachedFile(String uri) {
         URI u = URI.create(uri);
         String fullPath = StringUtils.defaultIfEmpty(u.getScheme(), StringUtils.EMPTY);
         System.out.println(u.getPath());
         String[] dirs = u.getPath().split("/");
-        for (String dir : dirs) {
-            fullPath += File.separator + dir;
+        for (int i = 0; i < dirs.length-1; ++i) {
+            fullPath += File.separator + dirs[i];
         }
         fullPath += File.separator + FilenameUtils.getName(uri);
         File effectiveDirectory = new File(cacheDirectory, fullPath);
@@ -82,14 +114,17 @@ public class URIResolver {
         int contentLength = 0;
         long lastModified = 0;
 
-        Reader in = new FileReader(metaFile);
-        Iterable<CSVRecord> records = CSVFormat.DEFAULT
-                .parse(in);
-        for (CSVRecord record : records) {
-            if (record.get(0).equals(cachedFile.getAbsolutePath())) {
-                contentLength = Integer.parseInt(record.get(1));
-                lastModified = Long.parseLong(record.get(2));
-                break;
+        BufferedReader in = new BufferedReader(new FileReader(metaFile));
+        String line = StringUtils.EMPTY;
+        while ((line = in.readLine()) != null) {
+            Iterable<CSVRecord> records = CSVFormat.DEFAULT
+                    .parse(new StringReader(line));
+            for (CSVRecord record : records) {
+                if (record.get(0).equals(cachedFile.getAbsolutePath())) {
+                    contentLength = Integer.parseInt(record.get(1));
+                    lastModified = Long.parseLong(record.get(2));
+                    break;
+                }
             }
         }
         if (lastModified > 0 && connection.getLastModified() == lastModified &&
@@ -98,24 +133,37 @@ public class URIResolver {
             return cachedFile;
         }
         else {
-            cachedFile.delete();
-            String metaData = cachedFile.getAbsolutePath() + "," + connection.getContentLength() + "," +
-                    connection.getLastModified();
-            FileUtils.writeStringToFile(metaFile, metaData, "UTF-8");
+            String warnOnly = StringUtils.defaultIfEmpty(System.getProperty("konduit.serving.cache.validation.warnonly"),"true");
+            if (warnOnly.equals("true")) {
+                log.error("Cached file " + cachedFile.getAbsolutePath() + " has inconsistent state.");
+                return cachedFile;
+            }
+            else {
+                log.error("Cached file " + cachedFile.getAbsolutePath() + " has inconsistent state and will be removed");
+                cachedFile.delete();
+            }
         }
+        // File was either just deleted or didn't exist, so writing metadata here and caching in
+        // the calling method.
+        String metaData = cachedFile.getAbsolutePath() + "," + connection.getContentLength() + "," +
+                connection.getLastModified() + "," + System.currentTimeMillis() + System.lineSeparator();
+        FileUtils.writeStringToFile(metaFile, metaData, "UTF-8", true);
         return null;
     }
 
     public static File getFile(String uri) throws IOException {
-
         URI u = URI.create(uri);
-        String scheme = u.getScheme();
-        if (scheme.equals("file")) {
-            return new File(u.getPath());
-        }
-        File cachedFile = getCachedFile(uri);
+        return getFile(u);
+    }
 
-        URL url = u.toURL();
+    public static File getFile(URI uri) throws IOException {
+        String scheme = uri.getScheme();
+        if (scheme.equals("file")) {
+            return new File(uri.getPath());
+        }
+        File cachedFile = getCachedFile(uri.getPath());
+
+        URL url = uri.toURL();
         if (cachedFile.exists()) {
             File verifiedFile = load(url, cachedFile);
             if (verifiedFile != null) {
